+++ src/sortuse64.c
@@ -274,8 +274,7 @@ SEXP r_ram_integer64_orderfin_asc(
 
   for(i=1;i<nt;i++)
     if (table[index[i]-1] < table[index[i-1]-1]) {
-      for(i=0; i<n; i++) ret[i] = FALSE;
-      goto wrapup_no_index_increment;
+      error("Invalid input -- 'table' is not sorted by 'order'");
     }
 
   for(i=0;i<nt;i++)
@@ -324,7 +323,6 @@ wrapup:
   for(i=0;i<nt;i++)
     index[i]++;
 
-wrapup_no_index_increment:
       R_Busy(0);
   if (method==0)
     error("unimplemented method");
@@ -361,8 +359,7 @@ SEXP r_ram_integer64_orderpos_asc(
 
   for(i=1;i<nt;i++)
     if (table[index[i]-1] < table[index[i-1]-1]) {
-      for(i=0; i<n; i++) ret[i] = nomatch;
-      goto wrapup_no_index_increment;
+      error("Invalid input -- 'table' is not sorted by 'order'");
     }
 
   for(i=0;i<nt;i++)
@@ -412,7 +409,6 @@ wrapup:
   for(i=0;i<nt;i++)
     index[i]++;
 
-wrapup_no_index_increment:
       R_Busy(0);
   if (method==0)
     error("unimplemented method");